+++ R/class_pipeline.R
@@ -147,11 +147,11 @@ pipeline_unload_transient <- function(pipeline) {
   }
 }
 
-pipeline_produce_subpipeline <- function(pipeline, name) {
+pipeline_produce_subpipeline <- function(pipeline, name, keep_value = NULL) {
   target <- pipeline_get_target(pipeline, name)
   deps <- target_deps_deep(target, pipeline)
   targets <- new.env(parent = emptyenv())
-  keep_value <- identical(target$settings$retrieval, "main")
+  keep_value <- keep_value %||% identical(target$settings$retrieval, "main")
   lapply(
     deps,
     pipeline_assign_target_copy,